+++ framework/model/src/main/java/is/codion/framework/model/EntityEditModel.java
@@ -126,12 +126,6 @@ public interface EntityEditModel {
 	EntitySearchModel searchModel(ForeignKey foreignKey);
 
 	/**
-	 * Refreshes the active Entity from the database, discarding all changes.
-	 * If the active Entity is new then calling this method has no effect.
-	 */
-	void refresh();
-
-	/**
 	 * Note: This method must be called on the UI thread in case a panel has been based on this model.
 	 * Performs an insert on the active entity, sets the primary key values of the active entity
 	 * according to the primary key of the inserted entity
@@ -437,6 +431,13 @@ public interface EntityEditModel {
 		 * Reverts all attribute value changes.
 		 */
 		void revert();
+
+		/**
+		 * Refreshes the active Entity from the database, discarding all changes.
+		 * If the active Entity is new then calling this method has no effect.
+		 * @see #exists()
+		 */
+		void refresh();
 
 		/**
 		 * @return an {@link ObservableState} indicating whether the entity exists in the database